move menu item processing into a service object
